Harden enrollment admission model APIs
