+++ apps/prairielearn/src/models/enrollment-reconciliation-concurrency.test.sql
@@ -1,10 +1,3 @@
--- BLOCK update_enrollment_pending_name
-UPDATE enrollments
-SET
-  pending_name = $pending_name
-WHERE
-  id = $enrollment_id;
-
 -- BLOCK insert_student_label
 INSERT INTO
   student_labels (course_instance_id, name, color, uuid)
@@ -29,15 +22,6 @@ WHERE
   enrollment_id = $enrollment_id
 ORDER BY
   student_label_id;
-
--- BLOCK select_student_label_enrollment_id
-SELECT
-  id
-FROM
-  student_label_enrollments
-WHERE
-  enrollment_id = $enrollment_id
-  AND student_label_id = $student_label_id;
 
 -- BLOCK insert_publishing_extension
 INSERT INTO
@@ -75,6 +59,14 @@ WHERE
   enrollment_id = $enrollment_id
   AND course_instance_publishing_extension_id = $publishing_extension_id;
 
+-- BLOCK select_publishing_extension_id
+SELECT
+  id
+FROM
+  course_instance_publishing_extensions
+WHERE
+  id = $publishing_extension_id;
+
 -- BLOCK insert_assessment_access_control_rule
 INSERT INTO
   assessment_access_control_rules (assessment_id, number, target_type, uuid)
@@ -109,10 +101,57 @@ WHERE
   enrollment_id = $enrollment_id
   AND assessment_access_control_rule_id = $rule_id;
 
--- BLOCK select_audit_event_by_id
+-- BLOCK select_audit_event_sequence_value
 SELECT
-  *
+  last_value::text
 FROM
-  audit_events
+  audit_events_id_seq;
+
+-- BLOCK lock_enrollment
+SELECT
+  id
+FROM
+  enrollments
 WHERE
-  id = $audit_event_id;
+  id = $enrollment_id
+FOR UPDATE;
+
+-- BLOCK lock_publishing_extension_enrollment
+SELECT
+  id
+FROM
+  course_instance_publishing_extension_enrollments
+WHERE
+  id = $id
+FOR UPDATE;
+
+-- BLOCK lock_assessment_access_control_enrollment
+SELECT
+  id
+FROM
+  assessment_access_control_enrollments
+WHERE
+  id = $id
+FOR UPDATE;
+
+-- BLOCK delete_enrollment
+DELETE FROM enrollments
+WHERE
+  id = $enrollment_id;
+
+-- BLOCK set_local_application_name
+SELECT
+  set_config('application_name', $application_name, true);
+
+-- BLOCK select_waiting_application_lock
+SELECT
+  1
+FROM
+  pg_stat_activity
+WHERE
+  pid <> pg_backend_pid()
+  AND application_name = $application_name
+  AND wait_event_type = 'Lock'
+  AND query LIKE $query_pattern
+LIMIT
+  1;
